+++ Runtime/Types/UIScreenView.cs
@@ -1,0 +1,10 @@
+namespace Gt.UI
+{
+    public abstract class UIScreenView : UIView
+    {
+        public virtual void OnNavigatedTo() { }
+
+        public virtual void OnNavigatedFrom() { }
+    }
+}
+